corrected the response status
create order will have the required paramaters to be a json with a formation like this:
{
    "orderLineItemsList": [
        {
            "id": 5,
            "product_name": "iphone_13",
            "quantity": 1
        },
        {
            "id": 6,
            "product_name": "jangmeyon",
            "quantity": 1
        }
    ],
    "customerId": "testing",
    "modeOfEating": "hi",
    "invoiceId": "None",
    "orderId": 3
}

note the names of each field in the json is impt. The placeOrder function will be implemented via this http://localhost:8081/api/order via a HTTP Post method.

for the getAllOrders function, it will be implemented via the url: http://localhost:8081/api/order via a HTTP Get method.

for the findOrderById function, it will be implemented via the url: http://localhost:8081/api/order/findOrderById. The parameter of this function will be OrderId which is a integer value. Returns code 400 if no order is found with the OrderId provided, else code 200 with the Order Json Object